requirements.txt

psutil
pypresence
requests
python-dotenv

# dev?
setuptools
flake8